Refactoring train.py pieces
