+++ MaxText/metric_logger.py
@@ -42,6 +42,26 @@ def _prepare_metrics_for_json(metrics, step, run_name):
   return metrics_dict
 
 
+def record_activation_metrics(output_metrics, intermediate_outputs, config):
+  """Adds the activation metrics to the metrics dict"""
+
+  if config.scan_layers:
+    metrics_dict = intermediate_outputs["intermediates"]["decoder"]["decoder"]
+
+    for layer_num in range(config.num_decoder_layers):
+      output_metrics["scalar"][f"activ_fraction_zero/layer_{layer_num:03d}"] = metrics_dict["activation_fraction_zero"][0][
+          layer_num
+      ]
+      output_metrics["scalar"][f"activ_mean/layer_{layer_num:03d}"] = metrics_dict["activation_mean"][0][layer_num]
+      output_metrics["scalar"][f"activ_stdev/layer_{layer_num:03d}"] = metrics_dict["activation_stdev"][0][layer_num]
+  else:
+    for layer_num in range(config.num_decoder_layers):
+      layer = intermediate_outputs["intermediates"]["decoder"][f"layers_{layer_num}"]
+      output_metrics["scalar"][f"activ_fraction_zero/layer_{layer_num:03d}"] = layer["activation_fraction_zero"][0]
+      output_metrics["scalar"][f"activ_mean/layer_{layer_num:03d}"] = layer["activation_mean"][0]
+      output_metrics["scalar"][f"activ_stdev/layer_{layer_num:03d}"] = layer["activation_stdev"][0]
+
+
 class MetricLogger:
   """
   Logger for saving metrics to a local file, GCS and TensorBoard.
