+++ MaxText/standalone_checkpointer.py
@@ -37,7 +37,8 @@ from MaxText import maxtext_utils
 from MaxText import train_utils
 from MaxText import max_logging
 from MaxText import pyconfig
-from MaxText.train import get_first_step, validate_train_config
+from MaxText.train import get_first_step
+from MaxText.train_utils import validate_train_config
 from MaxText.layers import models
 
 Transformer = models.Transformer
